pyproject.toml
[project]
name = "expense-bot"
version = "0.1.0"
description = "Telegram bot for expense tracking"
requires-python = ">=3.11"

dependencies = [
    "aiogram>=3.15",
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.20",
    "alembic>=1.14",
    "pydantic-settings>=2.7",
    "anthropic>=0.42",
    "mistralai>=1.0.0",
    "openpyxl>=3.1",
    "apscheduler>=3.10",
    "aiohttp>=3.11",
    "aiohttp-socks>=0.10.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]